added init files for bayes, refactored to base class
