+++ tuningdeap/__init__.py
@@ -1,1 +1,2 @@
-from tuningdeap.main import TuningDeap
+from tuningdeap.tuningdeap import TuningDeap
+from tuningdeap.tuningbayes import TuningBayes
